% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EfficientProMisesold.R
\name{EfficientProMisesold}
\alias{EfficientProMisesold}
\title{Efficient ProMises model}
\usage{
EfficientProMisesold(
  data,
  maxIt = 10,
  t = 0.001,
  k = 0,
  Q = NULL,
  ref_ds = NULL,
  scaling = T,
  reflection = T,
  subj = F,
  centered = T,
  coord = NULL
)
}
\arguments{
\item{data}{data, i.e., array of matrices with dimension time points - voxels or list of matrices with dimension time points - voxels}

\item{maxIt}{maximum number of iteration}

\item{t}{the threshold value to be reached as the minimum relative reduction between the mean matrices}

\item{k}{value of the concentration parameter of the prior distribution}

\item{Q}{value of the location parameter of the prior distribution. It has dimension time-points x time-points, it could be not symmetric.}

\item{ref_ds}{starting matrix to align}

\item{scaling}{Flag to apply scaling transformation}

\item{reflection}{Flag to apply reflection transformation}

\item{subj}{Flag if each subject has his/her own set of voxel after voxel selection step}

\item{centered}{center data?}

\item{coord}{3-dim or 2-dim coordinates of the variables. If \code{subj = F}  then coord is a matrix with dimensions voxels x 2/3, if \code{subj = T} then coord is a list of matrices 
with dimensions voxels x 2/3. If the location parameter \code{Q = NULL}, then \code{coord} is used to compute it}
}
\value{
\code{EfficientProMises} returns a list with four components:
\item{\code{Xest}}{an array with the aligned matrices}
\item{\code{R}}{an array with the rotation matrices}
\item{\code{dist}}{a vector with length equal to the number of iterations that contains the distances between a reference matrix and the previous one}
\item{\code{count}}{the number of iterations done by the algorithm}
}
\description{
Performs the functional alignment using the Efficient ProMises model. Matrices must have the same number of rows.
}
\references{
For the theory on the ProMises model see: A. Andreella and L. Finos
(2022), Procrustes analysis for high-dimensional data, 
Psychometrika 87, 1422-1438
}
\author{
Angela Andreella and Daniela Corbetta
}
