% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMFP.R
\name{vMFP}
\alias{vMFP}
\title{von Mises Fisher Procrustes model}
\usage{
vMFP(X, Q = NULL, k, ref_ds, scaling = TRUE, reflection = TRUE)
}
\arguments{
\item{X}{data, i.e., list of matrices with dimension time points - voxels}

\item{Q}{value of the location parameter of the prior distribution. It has dimension voxels x voxels, it could be not symmetric.}

\item{k}{value of the concentration parameter of the prior distribution}

\item{ref_ds}{starting matrix to align}

\item{scaling}{Flag to apply scaling transformation}

\item{reflection}{Flag to apply reflection transformation}
}
\value{
Returns list of matrices
}
\description{
perform functional alignment by von Mises Fisher Procrustes model
}
\author{
Angela Andreella
}
