\name{vMFPmodel-package}
\alias{vMFPmodel-package}
\alias{vMFPmodel}
\docType{package}
\title{
\packageTitle{vMFPmodel}
}
\description{
\packageDescription{vMFPmodel}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{vMFPmodel}
\packageIndices{vMFPmodel}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{vMFPmodel}

Maintainer: \packageMaintainer{vMFPmodel}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
