% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vMFPmodel.R
\name{vMFPmodel}
\alias{vMFPmodel}
\title{von Mises Fisher Procrustes model}
\usage{
vMFPmodel(data, maxIt=10, t=.001, k = 0, Q = NULL, ref_ds = NULL, scaling= T, reflection= T, subj= F)
}
\arguments{
\item{data}{data, i.e., array of matrices with dimension time points - voxels or list of matrices with dimension time points - voxels}

\item{maxIt}{maximum number of iteration}

\item{t}{the threshold value to be reached as the minimum relative reduction between the matrices}

\item{k}{value of the concentration parameter of the prior distribution}

\item{Q}{value of the location parameter of the prior distribution. It has dimension voxels x voxels, it could be not symmetric.}

\item{ref_ds}{starting matrix to align}

\item{scaling}{Flag to apply scaling transformation}

\item{reflection}{Flag to apply reflection transformation}

\item{subj}{Flag if each subject has his/her own set of voxel after voxel selection step}

\item{centered}{centered data?}
}
\value{
Returns list of matrices
}
\description{
perform the functional alignment using the von Mises Fisher Procrustes model
}
\author{
Angela Andreella and Daniela Corbetta
}
