% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EfficientProMises.R
\name{EfficientProMises}
\alias{EfficientProMises}
\title{Efficient ProMises model}
\usage{
EfficientProMises(data, maxIt=10, t=.001, k = 0, Q = NULL, ref_ds = NULL, scaling= T, reflection= T, subj= F, coord = NULL)
}
\arguments{
\item{data}{data, i.e., array of matrices with dimension time points - voxels or list of matrices with dimension time points - voxels}

\item{maxIt}{maximum number of iteration}

\item{t}{the threshold value to be reached as the minimum relative reduction between the matrices}

\item{k}{value of the concentration parameter of the prior distribution}

\item{Q}{value of the location parameter of the prior distribution. It has dimension voxels x voxels, it could be not symmetric.}

\item{ref_ds}{starting matrix to align}

\item{scaling}{Flag to apply scaling transformation}

\item{reflection}{Flag to apply reflection transformation}

\item{subj}{Flag if each subject has his/her own set of voxel after voxel selection step}

\item{centered}{center data?}

\item{coord}{3-dim coordinates of the variabiles}
}
\value{
\code{EfficientProMises} returns a list with four components:
\item{\code{Xest}}{an array with the aligned matrices}
\item{\code{R}}{an array with the rotation matrices}
\item{\code{dist}}{a vector with length equal to the number of iterations that contains the distances between a reference matrix and the previous one}
\item{\code{count}}{the number of iterations done by the algorithm}
}
\description{
Performs the functional alignment using the light version of the von Mises Fisher Procrustes model
}
\author{
Angela Andreella and Daniela Corbetta
}
