% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EfficientProMisesSubj.R
\name{EfficientProMisesSubj}
\alias{EfficientProMisesSubj}
\title{Efficient ProMises model}
\usage{
EfficientProMises(data, maxIt=10, t =.001, k = 0, Q = NULL, 
ref_ds = NULL, scaling = T, reflection= T, subj= T, coord = NULL)
}
\arguments{
\item{data}{data, i.e., list of matrices with dimension time points - voxels. Matrices can have different number of columns but must have same numbers of rows.}

\item{maxIt}{maximum number of iteration}

\item{t}{the threshold value to be reached as the minimum relative reduction between the matrices}

\item{k}{value of the concentration parameter of the prior distribution}

\item{Q}{value of the location parameter of the prior distribution. It has dimension time-points x time-points, it could be not symmetric.}

\item{ref_ds}{starting matrix to align}

\item{scaling}{Flag to apply scaling transformation}

\item{reflection}{Flag to apply reflection transformation}

\item{subj}{Flag if each subject has his/her own set of voxel after voxel selection step}

\item{coord}{list with 3-dim coordinates of the variables. Matrices can have different coordinates. If If the location parameter \code{Q = NULL}, then \code{coord} is used to compute it}
}
\value{
\code{EfficientProMises} returns a list with four components:
\item{\code{Xest}}{a list with the aligned matrices}
\item{\code{R}}{an array with the rotation matrices}
\item{\code{dist}}{a vector with length equal to the number of iterations that contains the distances between a reference matrix and the previous one}
\item{\code{count}}{the number of iterations done by the algorithm}
}
\description{
Performs the functional alignment using the light version of the von Mises Fisher Procrustes model allowing for different number of voxels between matrices
}
\author{
Angela Andreella and Daniela Corbetta
}
