% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kCalibrate.R
\name{kCalibrate}
\alias{kCalibrate}
\title{von Mises Fisher Procrustes model}
\usage{
kCalibrate(data)
}
\arguments{
\item{D}{squared euclidean distance matrix}

\item{p}{distances to compute bandwith value}
}
\value{
Returns list of matrices
}
\description{
perform the functional alignment using the von Mises Fisher Procrustes model
}
\author{
Angela Andreella and Daniela Corbetta
}
